+++ crates/omnigraph-api-types/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "omnigraph-api-types"
+version = "0.7.0"
+edition = "2024"
+description = "Shared HTTP wire DTOs for Omnigraph — request/response types and engine-result → DTO mappings used by both omnigraph-server and omnigraph-cli (RFC-009). Plain serde/utoipa types; no transport or server internals."
+license = "MIT"
+repository = "https://github.com/ModernRelay/omnigraph"
+homepage = "https://github.com/ModernRelay/omnigraph"
+documentation = "https://docs.rs/omnigraph-api-types"
+
+[dependencies]
+omnigraph = { package = "omnigraph-engine", path = "../omnigraph", version = "0.7.0" }
+omnigraph-compiler = { path = "../omnigraph-compiler", version = "0.7.0" }
+serde = { workspace = true }
+serde_json = { workspace = true }
+utoipa = { workspace = true }
