+++ crates/omnigraph-server/Cargo.toml
@@ -22,6 +22,7 @@ aws = ["dep:aws-config", "dep:aws-sdk-secretsmanager"]
 omnigraph = { package = "omnigraph-engine", path = "../omnigraph", version = "0.7.0" }
 omnigraph-compiler = { path = "../omnigraph-compiler", version = "0.7.0" }
 omnigraph-policy = { path = "../omnigraph-policy", version = "0.7.0" }
+omnigraph-api-types = { path = "../omnigraph-api-types", version = "0.7.0" }
 omnigraph-cluster = { path = "../omnigraph-cluster", version = "0.7.0" }
 axum = { workspace = true }
 clap = { workspace = true }
